Prevents logical race condition in migration
A final migrate (the swap of the old store for the new) can happen
before a writer locks the store but after it getting write permission
from the capacity watcher. At that point the writer will assume
enough storage in the store. However the new store might be at capacity
after migration.

This is fixed by checking for a running migration before doing a write
and by checking for a write in progress before doing migration.