[write/capacity] Regulate access to store per httpclient
Previously this was done by letting any httpclient/append through
but cancelling them inside the store if they are not writing at
the right pos.

To simplfy the store we now do this in the `wait_for_space` function.
On seek the httpclient gets a writer_token, only the stream with the
last writer_token can pass `wait_for_space`